requirements.txt
pandas
selenium
webdriver-manager
openpyxl
numpy
requests
python-dotenv